requirements.txt
dify_plugin
PyMuPDF
pillow
